# chore: add bound check for raft-engine logstore