--- a/src/mito2/src/region/opener.rs
+++ b/src/mito2/src/region/opener.rs
@@ -387,10 +387,13 @@ pub(crate) async fn replay_memtable<S: LogStore>(
     let mut region_write_ctx = RegionWriteCtx::new(region_id, version_control, wal_options.clone());
 
     let replay_from_entry_id = flushed_entry_id + 1;
+    let mut stale_entry_found = false;
     let mut wal_stream = wal.scan(region_id, replay_from_entry_id, wal_options)?;
     while let Some(res) = wal_stream.next().await {
         let (entry_id, entry) = res?;
         if entry_id <= flushed_entry_id {
+            stale_entry_found = true;
+            warn!("Stale WAL entries read during replay, region id: {}, flushed entry id: {}, entry id read: {}", region_id, flushed_entry_id, entry_id);
             ensure!(
                 allow_stale_entries,
                 StaleLogEntrySnafu {
@@ -399,7 +402,6 @@ pub(crate) async fn replay_memtable<S: LogStore>(
                     unexpected_entry_id: entry_id,
                 }
             );
-            warn!("Stale WAL entries read during replay, region id: {}, flushed entry id: {}, entry id read: {}", region_id, flushed_entry_id, entry_id);
         }
 
         last_entry_id = last_entry_id.max(entry_id);
@@ -417,7 +419,7 @@ pub(crate) async fn replay_memtable<S: LogStore>(
     region_write_ctx.set_next_entry_id(last_entry_id + 1);
     region_write_ctx.write_memtable();
 
-    if allow_stale_entries {
+    if allow_stale_entries && stale_entry_found {
         wal.obsolete(region_id, flushed_entry_id, wal_options)
             .await?;
         info!("Force obsolete WAL entries, region id: {}, flushed entry id: {}, last entry id read: {}", region_id, flushed_entry_id, last_entry_id);